Generic processTools extracted.
