+++ lib/github.js
@@ -1,6 +1,7 @@
 const axios = require('axios');
 const { generateToolsSystemPrompt, extractXmlTools, processXmlTools } = require('./xml-tools');
 const { getAndAuthorizeToken } = require('./auth');
+const { processTools } = require('./tools/processor');
 
 /**
  * Get available GitHub Copilot models
@@ -97,7 +98,7 @@ async function runPrompt(options) {
         // Extract and process XML tool blocks
         if (options.tools && options.tools.length > 0) {
           const toolBlocks = extractXmlTools(message.content);
-          await processXmlTools(toolBlocks);
+          await processTools(toolBlocks);
         }
       }
     }
