+++ lib/xml-tools.js
@@ -1,12 +1,5 @@
 const { XMLParser } = require('fast-xml-parser');
-
-const { fetch } = require('./tools/fetch');
-const { editor } = require('./tools/editor');
-
-const TOOLS_MAP = {
-  fetch: fetch,
-  editor: editor,
-};
+const { TOOLS_MAP } = require('./tools/tools-map');
 
 function processXmlCdata(xmlString) {
   let result = '';
@@ -112,36 +105,6 @@ function extractXmlTools(content) {
 }
 
 /**
- * Process and execute tools from extracted XML blocks
- * @param {Array<Object>} toolBlocks Array of tool blocks to process
- * @returns {Promise<void>}
- */
-async function processXmlTools(toolBlocks) {
-  toolBlocks.sort((a, b) => {
-    if (a.tool === b.tool) {
-      return TOOLS_MAP[a.tool].compare(a.params, b.params);
-    }
-    return 0;
-  });
-
-  console.log(JSON.stringify(toolBlocks, null, 2));
-
-  for (const block of toolBlocks) {
-    if (TOOLS_MAP[block.tool]) {
-      const tool = TOOLS_MAP[block.tool];
-      try {
-        const result = await tool.execute(block.params);
-        console.log(`Tool ${block.tool} executed:`, result);
-      } catch (error) {
-        console.error(`Error executing tool ${block.tool}:`, error.message);
-      }
-    } else {
-      console.error(`Unknown tool: ${block.tool}`);
-    }
-  }
-}
-
-/**
  * Generate system prompt with tool instructions
  * @param {Array<string>} toolNames Names of tools to include
  * @returns {string} The system prompt with tool instructions
@@ -169,6 +132,5 @@ function generateToolsSystemPrompt(toolNames) {
 
 module.exports = {
   extractXmlTools,
-  processXmlTools,
   generateToolsSystemPrompt,
 };
